+++ src/gam/settings/_gam.py
@@ -29,4 +29,5 @@ ADD_SOCIAL_SCORE = "⬆️"
 REMOVE_SOCIAL_SCORE = "⬇️"
 
 WAGER_ERROR_REACTION = "❗"
+WAGER_NO_MONEY_REACTION = "💸"
 WAGER_SUCCESS_REACTION = "✅"
